remove detected code smells #34
