+++ src/main/java/com/github/cameltooling/model/ComponentOptionModel.java
@@ -16,67 +16,13 @@
  */
 package com.github.cameltooling.model;
 
-public class ComponentOptionModel {
+public class ComponentOptionModel extends BaseOptionModel{
 
-    private String name;
-    private String kind;
-    private String group;
-    private String required;
-    private String type;
-    private String javaType;
     private String deprecated;
     private String secret;
     private String description;
     private String defaultValue;
     private String enums;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getKind() {
-        return kind;
-    }
-
-    public void setKind(String kind) {
-        this.kind = kind;
-    }
-
-    public String getGroup() {
-        return group;
-    }
-
-    public void setGroup(String group) {
-        this.group = group;
-    }
-
-    public String getRequired() {
-        return required;
-    }
-
-    public void setRequired(String required) {
-        this.required = required;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getJavaType() {
-        return javaType;
-    }
-
-    public void setJavaType(String javaType) {
-        this.javaType = javaType;
-    }
 
     public String getDeprecated() {
         return deprecated;
