+++ src/main/java/com/github/cameltooling/model/EndpointOptionModel.java
@@ -16,15 +16,9 @@
  */
 package com.github.cameltooling.model;
 
-public class EndpointOptionModel {
+public class EndpointOptionModel extends BaseOptionModel {
 
-    private String name;
-    private String kind;
-    private String group;
     private String label;
-    private String required;
-    private String type;
-    private String javaType;
     private String enums;
     private String prefix;
     private String multiValue;
@@ -34,29 +28,6 @@ public class EndpointOptionModel {
     private String description;
     private String enumValues;
 
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getKind() {
-        return kind;
-    }
-
-    public void setKind(String kind) {
-        this.kind = kind;
-    }
-
-    public String getGroup() {
-        return group;
-    }
-
-    public void setGroup(String group) {
-        this.group = group;
-    }
 
     public String getLabel() {
         return label;
@@ -64,30 +35,6 @@ public class EndpointOptionModel {
 
     public void setLabel(String label) {
         this.label = label;
-    }
-
-    public String getRequired() {
-        return required;
-    }
-
-    public void setRequired(String required) {
-        this.required = required;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getJavaType() {
-        return javaType;
-    }
-
-    public void setJavaType(String javaType) {
-        this.javaType = javaType;
     }
 
     public String getEnums() {
@@ -153,5 +100,4 @@ public class EndpointOptionModel {
     public void setEnumValues(String enumValues) {
         this.enumValues = enumValues;
     }
-
 }
